First step: put sdl code in his file and namespace.
